fix "case value not in enumerated type" compiler warning
the proper fix would be if the driver headers defined the correct
enumerations, but until this happens, this will do

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@1441 e54a6e83-5905-42d5-8d5c-058d10e6a962

+++ src/zapit/src/frontend.cpp
@@ -595,7 +595,9 @@ int CFrontend::setFrontend(const struct dvb_frontend_parameters *feparams, bool 
 		return 0;
 	}
 
-	switch (fec_inner) {
+	/* the ugly cast avoids a compiler warning, because the Coolstream
+	 * private values are *not* fe_code_rate_t values */
+	switch ((int)fec_inner) {
 	case FEC_1_2:
 	case FEC_S2_QPSK_1_2:
 	case FEC_S2_8PSK_1_2:
